+++ src/main/java/org/nrg/containers/model/command/entity/CommandWrapperDerivedInputEntity.java
@@ -32,6 +32,7 @@ public class CommandWrapperDerivedInputEntity {
     private String rawReplacementKey;
     private boolean required = false;
     private boolean loadChildren = true;
+    private Boolean sensitive;
 
     private String derivedFromWrapperInput;
     private String derivedFromXnatObjectProperty;
@@ -59,6 +60,7 @@ public class CommandWrapperDerivedInputEntity {
         this.setRawReplacementKey(commandWrapperInput.rawReplacementKey());
         this.setRequired(commandWrapperInput.required());
         this.setLoadChildren(commandWrapperInput.loadChildren());
+        this.setSensitive(commandWrapperInput.sensitive());
         switch (commandWrapperInput.type()) {
             case "string":
                 this.setType(CommandWrapperInputType.STRING);
@@ -244,6 +246,13 @@ public class CommandWrapperDerivedInputEntity {
         this.loadChildren = loadChildren;
     }
 
+    public Boolean getSensitive() {
+        return sensitive;
+    }
+
+    public void setSensitive(final Boolean sensitive) {
+        this.sensitive = sensitive;
+    }
 
     @Override
     public boolean equals(final Object o) {
@@ -277,6 +286,7 @@ public class CommandWrapperDerivedInputEntity {
                 .add("rawReplacementKey", rawReplacementKey)
                 .add("required", required)
                 .add("loadChildren", loadChildren)
+                .add("sensitive", sensitive)
                 .toString();
     }
 }