+++ src/main/java/org/nrg/containers/model/command/entity/CommandWrapperExternalInputEntity.java
@@ -32,6 +32,7 @@ public class CommandWrapperExternalInputEntity {
     private String rawReplacementKey;
     private boolean required = false;
     private boolean loadChildren = true;
+    private Boolean sensitive;
 
     @Nonnull
     public static CommandWrapperExternalInputEntity fromPojo(final @Nonnull Command.CommandWrapperInput commandWrapperInput) {
@@ -54,6 +55,7 @@ public class CommandWrapperExternalInputEntity {
         this.setRawReplacementKey(commandWrapperInput.rawReplacementKey());
         this.setRequired(commandWrapperInput.required());
         this.setLoadChildren(commandWrapperInput.loadChildren());
+        this.setSensitive(commandWrapperInput.sensitive());
         switch (commandWrapperInput.type()) {
             case "string":
                 this.setType(CommandWrapperInputType.STRING);
@@ -221,6 +223,14 @@ public class CommandWrapperExternalInputEntity {
         this.loadChildren = loadChildren;
     }
 
+    public Boolean getSensitive() {
+        return sensitive;
+    }
+
+    public void setSensitive(final Boolean sensitive) {
+        this.sensitive = sensitive;
+    }
+
     @Override
     public boolean equals(final Object o) {
         if (this == o) return true;
@@ -251,6 +261,7 @@ public class CommandWrapperExternalInputEntity {
                 .add("rawReplacementKey", rawReplacementKey)
                 .add("required", required)
                 .add("loadChildren", loadChildren)
+                .add("sensitive", sensitive)
                 .toString();
     }
 }